+++ Quran.WindowsPhone/UI/ArabicFontSizeConverter.cs
@@ -18,10 +18,10 @@ namespace Quran.WindowsPhone.UI
             if (value != null && value.ToString() == "TranslationViewHeader")
                 return 50;
             else if (value != null && value.ToString() == "ArabicText")
-                return (SettingsUtils.Get<int>(Constants.PREF_TRANSLATION_TEXT_SIZE) *
+                return (SettingsUtils.Get<double>(Constants.PREF_TRANSLATION_TEXT_SIZE) *
                         Constants.ARABIC_FONT_SCALE_RELATIVE_TO_TRANSLATION);
             else
-                return SettingsUtils.Get<int>(Constants.PREF_TRANSLATION_TEXT_SIZE);
+                return SettingsUtils.Get<double>(Constants.PREF_TRANSLATION_TEXT_SIZE);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
